requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
email-validator==2.1.1
qrcode==7.4.2
pillow==10.2.0
pyzbar==0.1.9
python-multipart==0.0.9
